move tests that fail on all platforms out of the way
