+++ test/integration/component/test_ps_limits.py
@@ -506,7 +506,8 @@ class TestVolumeLimits(cloudstackTestCase):
         return
 
     @data(ROOT_DOMAIN_ADMIN, CHILD_DOMAIN_ADMIN)
-    @attr(tags=["advanced", "basic"], required_hardware="true")
+    # @attr(tags=["advanced", "basic"], required_hardware="true")
+    @attr(tags=["TODO"], required_hardware="true")
     def test_create_template_snapshot(self, value):
         """Test create snapshot and templates from volume
