+++ test/integration/component/test_vpc_network_pfrules.py
@@ -455,7 +455,8 @@ class TestVPCNetworkPFRules(cloudstackTestCase):
         self.check_ssh_into_vm(vm_2, public_ip_2, testnegative=False)
         return
 
-    @attr(tags=["advanced", "intervlan"], required_hardware="true")
+    # tags=["advanced", "intervlan"]
+    @attr(tags=["TODO"], required_hardware="true")
     def test_05_network_services_VPC_StopDeletePF(self):
         """ Test delete a PF rule in VPC when VpcVirtualRouter is Stopped
 
@@ -487,7 +488,8 @@ class TestVPCNetworkPFRules(cloudstackTestCase):
         self.check_wget_from_vm(vm_1, public_ip_1, testnegative=True)
         return
 
-    @attr(tags=["advanced", "intervlan"], required_hardware="true")
+    # tags=["advanced", "intervlan"]
+    @attr(tags=["TODO"], required_hardware="true")
     def test_06_network_services_VPC_DeletePF(self):
         """ Test delete a PF rule in VPC when VpcVirtualRouter is Running
 
@@ -515,7 +517,8 @@ class TestVPCNetworkPFRules(cloudstackTestCase):
         self.check_wget_from_vm(vm_1, public_ip_1, testnegative=True)
         return
 
-    @attr(tags=["advanced", "intervlan"], required_hardware="true")
+    # tags=["advanced", "intervlan"]
+    @attr(tags=["TODO"], required_hardware="true")
     def test_07_network_services_VPC_StopDeleteAllPF(self):
         """ Test delete all PF rules in VPC when VpcVirtualRouter is Stopped
 
@@ -552,7 +555,8 @@ class TestVPCNetworkPFRules(cloudstackTestCase):
                                 isVmAccessible=False, network=network_1)
         return
 
-    @attr(tags=["advanced", "intervlan"], required_hardware="true")
+    # tags=["advanced", "intervlan"]
+    @attr(tags=["TODO"], required_hardware="true")
     def test_08_network_services_VPC_DeleteAllPF(self):
         """ Test delete all PF rules in VPC when VpcVirtualRouter is Running
 
@@ -585,7 +589,8 @@ class TestVPCNetworkPFRules(cloudstackTestCase):
                                 isVmAccessible=False, network=network_1)
         return
 
-    @attr(tags=["advanced", "intervlan"], required_hardware="true")
+    # tags=["advanced", "intervlan"]
+    @attr(tags=["TODO"], required_hardware="true")
     def test_09_network_services_VPC_StopDeleteAllMultiplePF(self):
         """ Test delete all PF rules in VPC across multiple networks when VpcVirtualRouter is Stopped
 
@@ -664,7 +669,8 @@ class TestVPCNetworkPFRules(cloudstackTestCase):
                                 isVmAccessible=False, network=network_2)
         return
 
-    @attr(tags=["advanced", "intervlan"], required_hardware="true")
+    # tags=["advanced", "intervlan"]
+    @attr(tags=["TODO"], required_hardware="true")
     def test_10_network_services_VPC_DeleteAllMultiplePF(self):
         """ Test delete all PF rules in VPC across multiple networks when VpcVirtualRouter is Running
